moving to more weak dependencies
