+++ NEWS.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] – unreleased
+
+### Changed
+
+* `Distributions.jl` and `RecursiveArrayTools.jl` were moved to weak dependencies to reduce load time and improve extensibility.
+
 ## [0.9.20] – 2024-06-17
 
 ### Added
