+++ ext/ManifoldsDistributionsExt/projected_distribution.jl
