--- src/product_representations.jl
@@ -1,1 +1,0 @@
-
